+++ lib/constant.dart
@@ -1,1 +1,2 @@
-const double percentage = 1.11; // This is for cell size
+/// This is for cell size
+const double percentage = 1.11;