+++ lib/models/body_settings.dart
@@ -1,11 +1,19 @@
 import 'package:flutter/material.dart';
 
-//This contains calendar body settings to customize body color
+///This contains calendar body settings to customize body color
 class BodySettings {
-  final Color titleBGColor; // to customize body color
-  final Color backgroundColor; //default value is grey
-  final double? radius; //customize radius of body cells
+  final Color titleBGColor;
+
+  /// to customize body color
+  final Color backgroundColor;
+
+  ///default value is grey
+  final double? radius;
+
+  ///customize radius of body cells
   final double height;
+
+  /// customizable body-cell height
 
   BodySettings(
       {this.titleBGColor = Colors.grey,